moved dimensions endpoints and test into separate files, updated make file to allow private endpoint config to be overwritten by env config if set
