+++ Makefile
@@ -7,7 +7,7 @@ BIN_DIR?=.
 export GOOS?=$(shell go env GOOS)
 export GOARCH?=$(shell go env GOARCH)
 
-export ENABLE_PRIVATE_ENDPOINTS=true
+export ENABLE_PRIVATE_ENDPOINTS?=true
 
 build:
 	@mkdir -p $(BUILD_ARCH)/$(BIN_DIR)
